Move settings-related components to components/settings
