+++ client/extensions/wp-job-manager/components/settings/job-listings/style.scss
